[org] move hash out of hash_dot file
